requirements.txt
numpy==1.15.2
scikit-image
torch
torchvision
pillow==8.1.1
opencv-python
paddlepaddle
paddlehub
gradio
